requirements/release.txt
blinker
click
croniter
pytest
pytest-cov
python-daemon
redis
